% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb-organismdb.r
\docType{package}
\name{eupathdb-organismdb}
\alias{eupathdb-organismdb}
\alias{eupathdb-organismdb-package}
\title{eupathdb-organismdb: Make annotation packages for Tryps easier!}
\description{
This should bring together the annotation data at the TriTrypDB, gene ontology data, and KEGG.
Soon other datatypes should follow, especially the COG data.
}

